Implement GitHub label auto-creation and add new dependencies
Added a new method `ensure_labels_exist` to `github_issues.rb` that checks if required labels exist in the target repository using the `gh` CLI and creates them if they do not. This method is called before creating an issue to ensure labels are available.

Added the following gems to the Gemfile:
- `dotenv`
- `google-apis-customsearch_v1`
- `googleauth`
- `os`
- `tty-option`

Updated the logging require path in `github_issues.rb` from `memexrag/logging` to `logging`.

Signed-off-by: Robert Pannick <rwpannick@gmail.com>